+++ source/Internal/Service/UserReviewAndRatingServiceInterface.php
@@ -16,6 +16,15 @@ use Doctrine\Common\Collections\ArrayCollection;
 interface UserReviewAndRatingServiceInterface
 {
     /**
+     * Get number of reviews by given user.
+     *
+     * @param string $userId
+     *
+     * @return int
+     */
+    public function getReviewAndRatingListCount($userId);
+
+    /**
      * Returns Collection of User Ratings and Reviews.
      *
      * @param string $userId